[REFACTOR] 순환참조 해결
